:art: Moves files around and renames them for consistency
