+++ lib/core/lg-logic.js
