+++ lib/core/lg-utilities.js
