+++ lib/lost-column.js
@@ -1,7 +1,7 @@
 var newBlock = require('./new-block.js');
 
-var lgLogic = require('./_lg-logic');
-var lgUtils = require('./_lg-utilities');
+var lgLogic = require('./core/lg-logic');
+var lgUtils = require('./core/lg-utilities');
 
 module.exports = function lostColumnDecl(css, settings, result) {
   css.walkDecls('lost-column', function lostColumnFunction(decl) {
