+++ lib/lost-gutter.js
