+++ lib/lost-masonry-column.js
@@ -1,4 +1,4 @@
-var lgUtils = require('./_lg-utilities');
+var lgUtils = require('./core/lg-utilities');
 
 module.exports = function lostMasonryColumnDecl(css, settings) {
   css.walkDecls('lost-masonry-column', function lostMasonryColumnFunction(
