+++ lib/lost-move.js
@@ -1,4 +1,4 @@
-var lgUtils = require('./_lg-utilities');
+var lgUtils = require('./core/lg-utilities');
 
 module.exports = function lostMoveDecl(css, settings) {
   css.walkDecls('lost-move', function lostMoveDeclFunction(decl) {
