+++ lib/lost-offset.js
@@ -1,4 +1,4 @@
-var lgUtils = require('./_lg-utilities');
+var lgUtils = require('./core/lg-utilities');
 
 module.exports = function lostOffsetDecl(css, settings) {
   css.walkDecls('lost-offset', function lostOffsetDeclFunction(decl) {
