+++ lib/lost-row.js
@@ -1,7 +1,7 @@
 var newBlock = require('./new-block.js');
 
-var lgLogic = require('./_lg-logic');
-var lgUtils = require('./_lg-utilities');
+var lgLogic = require('./core/lg-logic');
+var lgUtils = require('./core/lg-utilities');
 
 module.exports = function lostRowDecl(css, settings, result) {
   css.walkDecls('lost-row', function lostRowDeclFunction(decl) {
