+++ lib/lost-utility.js
@@ -1,5 +1,5 @@
 var newBlock = require('./new-block.js');
-var getColorValue = require('./_lg-utilities').getColorValue;
+var getColorValue = require('./core/lg-utilities').getColorValue;
 
 function unitsMatch() {
   var args = Array.prototype.slice.call(arguments, 0);
