+++ lost.js
@@ -14,14 +14,14 @@ var lostOffset = require('./lib/lost-offset');
 var lostMove = require('./lib/lost-move');
 var lostMasonryWrap = require('./lib/lost-masonry-wrap');
 var lostMasonryColumn = require('./lib/lost-masonry-column');
-var lgGutter = require('./lib/lg-gutter');
+var lostGutter = require('./lib/lost-gutter');
 var lostVars = require('./lib/lost-vars');
 
 // Lost At Rules and Declarations
 var libs = [
   lostAtRule,
   lostVars,
-  lgGutter,
+  lostGutter,
   lostMove,
   lostUtility,
   lostFlexContainer,
